Rename new Epoch type
